+++ src/Components/Aside Right/index.jsx
@@ -1,4 +1,18 @@
-export const AsideRight = () => {
+import { useState } from "react";
+import { useEffect } from "react";
+import {getAllPosts} from "../../Components/Api/api"
+
+export const AsideRight = (props) => {
+    const [postsList, setPostsList] = useState([]);
+
+    useEffect(() => {
+      const getPosts = async () => {
+        const data = await getAllPosts();
+        setPostsList(data);
+      };
+      getPosts();
+    }, []);
+
     return(
         <>
             <h1>Right</h1>
