+++ src/Pages/App/App.jsx
@@ -1,6 +1,4 @@
 import { useRoutes, BrowserRouter } from "react-router-dom";
-import { useState } from "react";
-import { useEffect } from "react";
 import { Home } from "../Home";
 import { SignIn } from "../SignIn";
 import { NewPost } from "../NewPost";
@@ -18,27 +16,6 @@ const AppRoutes = () => {
 };
 
 function App() {
-  const [postsList, setPostsList] = useState([]);
-
-  useEffect(() => {
-    const getUsers = async () => {
-      const response = await fetch(
-        "https://javascript-challenge-f0392-default-rtdb.firebaseio.com/posts.json"
-      );
-      const data = await response.json();
-      console.log(data);
-      /* ["user1","user2"]*/ /*data[key] se convierte en {name:"...", picture:"..."}*/
-      const formattedData = Object.keys(data).map((key) => ({
-        key,
-        ...data[key],
-      }));
-      console.log(formattedData);
-      setPostsList(formattedData);
-    };
-    getUsers();
-  }, []);
-
-  console.log(postsList);
   return (
     <BrowserRouter>
       <AppRoutes />
